images/solution/requirements.txt
mysqlclient ~= 1.4
SQLAlchemy ~= 1.3
pandas
numpy
pymysql
mysql.connector
